works I2c

+++ Fpga.cpp
@@ -45,18 +45,19 @@ void Fpga::find_reg_address(uint32_t register_address)
 void Fpga::read(uint32_t* data)
 {
     *data = *( ( uint32_t *) map_register_address );
-  //  printf( "Data = 0x%08x  stays in the 0x%08lx\n ", data, register_address);     
+    
 }
 
 void Fpga::read(uint8_t* data)
 {
     *data = *( ( uint32_t *) map_register_address );
-  //  printf( "Data = 0x%08x  stays in the 0x%08lx\n ", data, register_address);     
 }
+
 
 void Fpga::read_fpga(uint32_t register_address,uint32_t* data)
 {
     this->find_reg_address(register_address);
+    //std::cout << "Reg address = " << std::hex << unsigned(register_address) << std::endl;
     this->read(data);
 }
 
@@ -73,7 +74,7 @@ void Fpga::write(uint32_t data)
 
 void Fpga::write(uint8_t data)
 {
-    *( ( uint32_t *) map_register_address ) = data;
+    *( ( uint8_t *) map_register_address ) = data;
 }
 
 void Fpga::write_fpga(uint32_t register_address,uint32_t data)
